requirements.txt
python-datamuse==1.2.1
nltk==3.4
flask==1.0.2
Flask-Cors==3.0.7
